Moved types for reading python JSON files from robotpy_data.ts to python_json_types.ts. Moved function initialize from robotpy_initialize.ts to python.ts.
Update generated json file for robotpy.
Add generated json file for external samples.
