+++ src/blocks/utils/robotpy_data.ts
@@ -19,86 +19,7 @@
  * @author lizlooney@google.com (Liz Looney)
  */
 
+import { PythonData } from './python_json_types';
 import generatedRobotPyData from './generated/robotpy_data.json';
 
-export const robotPyData = generatedRobotPyData as RobotPyData;
-
-export class RobotPyData {
-  modules: ModuleData[] = [];
-  classes: ClassData[] = [];
-  aliases: {[key: string]: string} = {};
-  subclasses: {[key: string]: string[]} = {};
-}
-
-export class ModuleData {
-  moduleName: string = '';
-  moduleVariables: VarData[] = [];
-  functions: FunctionData[] = [];
-  enums: EnumData[] = [];
-}
-
-export class ClassData {
-  className: string = '';
-  moduleName: string = '';
-  classVariables: VarData[] = [];
-  instanceVariables: VarData[] = [];
-  constructors: FunctionData[] = [];
-  instanceMethods: FunctionData[] = [];
-  staticMethods: FunctionData[] = [];
-  enums: EnumData[] = [];
-}
-
-export class VarData {
-  name: string = '';
-  type: string = '';
-  writable: boolean = false;
-  tooltip: string = '';
-}
-
-export class FunctionData {
-  functionName: string = '';
-  tooltip: string = '';
-  returnType: string = '';
-  args: ArgData[] = [];
-  declaringClassName?: string = '';
-}
-
-export class ArgData {
-  name: string = '';
-  type: string = '';
-  defaultValue: string = '';
-}
-
-export class EnumData {
-  enumClassName: string = '';
-  moduleName: string = '';
-  enumValues: string[] = [];    
-  tooltip: string = '';
-}
-
-export class VariableGettersAndSetters {
-  varNamesForGetter: string[] = [];
-  tooltipsForGetter: string[] = [];
-  varNamesForSetter: string[] = [];
-  tooltipsForSetter: string[] = [];
-}
-
-export function organizeVarDataByType(vars: VarData[]): {[key: string]: VariableGettersAndSetters} {
-  const varsByType: {[key: string]: VariableGettersAndSetters} = {}
-  for (const varData of vars) {
-    let variableGettersAndSetters: VariableGettersAndSetters;
-    if (varData.type in varsByType) {
-       variableGettersAndSetters = varsByType[varData.type];
-    } else {
-       variableGettersAndSetters = new VariableGettersAndSetters();
-       varsByType[varData.type] = variableGettersAndSetters;
-    }
-    variableGettersAndSetters.varNamesForGetter.push(varData.name);
-    variableGettersAndSetters.tooltipsForGetter.push(varData.tooltip);
-    if (varData.writable) {
-      variableGettersAndSetters.varNamesForSetter.push(varData.name);
-      variableGettersAndSetters.tooltipsForSetter.push(varData.tooltip);
-    }
-  }
-  return varsByType;
-}
+export const robotPyData = generatedRobotPyData as PythonData;
