+++ src/editor/editor.ts
@@ -25,6 +25,7 @@ import { extendedPythonGenerator } from './extended_python_generator';
 import { GeneratorContext } from './generator_context';
 import * as commonStorage from '../storage/common_storage';
 import { getToolboxJSON } from '../toolbox/toolbox';
+//import { testAllBlocksInToolbox } from '../toolbox/toolbox_tests';
 import { MethodsCategory} from '../toolbox/methods_category';
 
 
@@ -165,6 +166,7 @@ export class Editor {
     if (toolbox != this.toolbox) {
       this.toolbox = toolbox;
       this.blocklyWorkspace.updateToolbox(toolbox);
+      //testAllBlocksInToolbox();
     }
   }
 
@@ -194,9 +196,9 @@ export class Editor {
         }, 50);
         return;
       }
-      const exportedBlocks = commonStorage.extractExportedBlocks(
+      const robotBlocks = commonStorage.extractExportedBlocks(
          this.currentModule.projectName, this.projectContent);
-      this.setToolbox(getToolboxJSON(exportedBlocks, shownPythonToolboxCategories));
+      this.setToolbox(getToolboxJSON(robotBlocks, shownPythonToolboxCategories));
     }
   }
