+++ src/toolbox/toolbox.ts
@@ -31,15 +31,15 @@ import {category as methodsCategory} from './methods_category';
 import {category as componentSampleCategory} from './component_samples_category';
 
 export function getToolboxJSON(
-    opt_includeExportedBlocksFromProject: toolboxItems.ContentsType[],
+    opt_robotBlocks: toolboxItems.ContentsType[],
     shownPythonToolboxCategories: Set<string> | null) {
   const contents: toolboxItems.ContentsType[] = [];
 
   const robotPyCategories: toolboxItems.ContentsType[] = robotPyToolbox.getToolboxCategories();
-  filterGeneratedCategories(robotPyCategories, shownPythonToolboxCategories);
+  filterRobotPyCategories(robotPyCategories, shownPythonToolboxCategories);
   contents.push.apply(contents, robotPyCategories);
 
-  if (opt_includeExportedBlocksFromProject.length) {
+  if (opt_robotBlocks.length) {
     contents.push.apply(
       contents,
       [
@@ -49,7 +49,7 @@ export function getToolboxJSON(
         {
           kind: 'category',
           name: 'Project',
-          contents: opt_includeExportedBlocksFromProject,
+          contents: opt_robotBlocks,
         },
       ]);
   }
@@ -85,14 +85,14 @@ export function getToolboxJSON(
   };
 }
 
-function filterGeneratedCategories(
+function filterRobotPyCategories(
     contents: toolboxItems.ContentsType[], shownPythonToolboxCategories: Set<string> | null) {
   contents.forEach((item) => {
     if (item.kind === 'category') {
       const category = item as toolboxItems.Category;
       // Traverse the tree depth-first so we can easily identify and remove empty categories.
       if (category.contents) {
-        filterGeneratedCategories(category.contents, shownPythonToolboxCategories);
+        filterRobotPyCategories(category.contents, shownPythonToolboxCategories);
       }
       if ((category as toolboxItems.PythonModuleCategory).moduleName) {
         const moduleName = (item as toolboxItems.PythonModuleCategory).moduleName;
